+++ osu.Game/Screens/Play/PlayerSettings/VisualSettings.cs
@@ -15,7 +15,6 @@ namespace osu.Game.Screens.Play.PlayerSettings
         private readonly PlayerSliderBar<double> dimSliderBar;
         private readonly PlayerSliderBar<double> blurSliderBar;
         private readonly PlayerCheckbox showStoryboardToggle;
-        private readonly PlayerCheckbox showVideoToggle;
         private readonly PlayerCheckbox beatmapSkinsToggle;
         private readonly PlayerCheckbox beatmapHitsoundsToggle;
 
@@ -43,8 +42,7 @@ namespace osu.Game.Screens.Play.PlayerSettings
                 {
                     Text = "切换:"
                 },
-                showStoryboardToggle = new PlayerCheckbox { LabelText = "故事板" },
-                showVideoToggle = new PlayerCheckbox { LabelText = "背景视频" },
+                showStoryboardToggle = new PlayerCheckbox { LabelText = "故事板 / 背景视频" },
                 beatmapSkinsToggle = new PlayerCheckbox { LabelText = "谱面皮肤" },
                 beatmapHitsoundsToggle = new PlayerCheckbox { LabelText = "谱面击打音效" }
             };
@@ -56,7 +54,6 @@ namespace osu.Game.Screens.Play.PlayerSettings
             dimSliderBar.Bindable = config.GetBindable<double>(OsuSetting.DimLevel);
             blurSliderBar.Bindable = config.GetBindable<double>(OsuSetting.BlurLevel);
             showStoryboardToggle.Current = config.GetBindable<bool>(OsuSetting.ShowStoryboard);
-            showVideoToggle.Current = config.GetBindable<bool>(OsuSetting.ShowVideoBackground);
             beatmapSkinsToggle.Current = config.GetBindable<bool>(OsuSetting.BeatmapSkins);
             beatmapHitsoundsToggle.Current = config.GetBindable<bool>(OsuSetting.BeatmapHitsounds);
         }