+++ Changelog.md
@@ -219,3 +219,4 @@
 ### 2020/3/25
 *   合并上游pr
     *   [m] [Allow individual storyboard layers to disable masking #8426]
+    *   [m] [Start background video playback based on provided offset #8167]